refactor: extract sandbox path safety
